# Optimize District Ranking Query Performance