--- a/src/inngest/functions/districtsRankings/updateRankings.ts
+++ b/src/inngest/functions/districtsRankings/updateRankings.ts
@@ -1,11 +1,7 @@
-import { Logger } from 'inngest/middleware/logger'
-import pRetry from 'p-retry'
-
 import { inngest } from '@/inngest/inngest'
 import { onScriptFailure } from '@/inngest/onScriptFailure'
 import { REDIS_KEYS } from '@/utils/server/districtRankings/constants'
 import {
-  AdvocatesCountResult,
   getAdvocatesCountByDistrict,
   getReferralsCountByDistrict,
 } from '@/utils/server/districtRankings/getRankingData'
@@ -18,6 +14,7 @@ import { USStateCode } from '@/utils/shared/stateMappings/usStateUtils'
 const UPDATE_DISTRICT_RANKINGS_CRON_JOB_FUNCTION_ID = 'script.update-districts-rankings'
 const UPDATE_DISTRICT_RANKINGS_CRON_JOB_SCHEDULE = '0 */1 * * *' // every 1 hour
 const UPDATE_DISTRICT_RANKINGS_INNGEST_EVENT_NAME = 'script/update-districts-rankings'
+const DB_CONNECTION_LIMIT = 50
 
 export interface UpdateDistrictsRankingsCronJobSchema {
   name: typeof UPDATE_DISTRICT_RANKINGS_INNGEST_EVENT_NAME
@@ -27,6 +24,7 @@ export const updateDistrictsRankings = inngest.createFunction(
   {
     id: UPDATE_DISTRICT_RANKINGS_CRON_JOB_FUNCTION_ID,
     onFailure: onScriptFailure,
+    concurrency: DB_CONNECTION_LIMIT,
   },
   {
     ...(NEXT_PUBLIC_ENVIRONMENT === 'production'
@@ -39,40 +37,44 @@ export const updateDistrictsRankings = inngest.createFunction(
       return await syncReferralsWithoutAddress()
     })
 
-    const [districtAdvocatesCounts, districtsReferralsCount] = await Promise.all([
-      step.run('Get Advocates Count by District', async () => {
-        logger.info('Getting Advocates Count by District')
-        return await fetchStatesCount({
-          queryFn: getAdvocatesCountByDistrict,
-          logger,
-        })
-      }),
-      step.run('Get Referrals Count by District', async () => {
-        logger.info('Getting Referrals Count by District')
-        return await fetchStatesCount({
-          queryFn: getReferralsCountByDistrict,
-          logger,
-        })
-      }),
-    ])
+    const stateCodes = Object.keys(US_STATE_CODE_TO_DISTRICT_COUNT_MAP) as USStateCode[]
+    const advocatesCountPromises = []
+    const referralsCountPromises = []
+    for (const stateCode of stateCodes) {
+      advocatesCountPromises.push(
+        step.run(`Get Advocates Count by District for ${stateCode}`, async () => {
+          logger.info(`Getting Advocates Count by District for ${stateCode}`)
+          return await getAdvocatesCountByDistrict(stateCode)
+        }),
+      )
+      referralsCountPromises.push(
+        step.run(`Get Referrals Count by District for ${stateCode}`, async () => {
+          logger.info(`Getting Referrals Count by District for ${stateCode}`)
+          return await getReferralsCountByDistrict(stateCode)
+        }),
+      )
+    }
+    const advocatesCountsByDistrictEntries = (await Promise.all(advocatesCountPromises)).flat()
+    const referralsCountByDistrictEntries = (await Promise.all(referralsCountPromises)).flat()
 
-    const [districtAdvocatesRankingResults, districtReferralsRankingResults] = await Promise.all([
+    const [advocatesRankingResults, referralsRankingResults] = await Promise.all([
       step.run('Update Districts Advocates Rankings', async () => {
         const upsertDistrictAdvocatesRanking = await createDistrictRankingUpserter(
           REDIS_KEYS.DISTRICT_ADVOCATES_RANKING,
         )
         const results = await Promise.all(
-          districtAdvocatesCounts.map(entry => upsertDistrictAdvocatesRanking(entry)),
+          advocatesCountsByDistrictEntries.map(entry => upsertDistrictAdvocatesRanking(entry)),
         )
 
         const successfulResults = results.filter(result => result.success)
         const failedResults = results.filter(result => !result.success)
 
         return {
-          totalResults: results.length,
-          successfulResults: successfulResults.length,
-          failedResults: failedResults.length,
+          totalEntries: results.length,
+          successfulEntries: successfulResults.map(result => result.entry),
+          successful: successfulResults.length,
           failedEntries: failedResults.map(result => result.entry),
+          failed: failedResults.length,
         }
       }),
 
@@ -81,62 +83,27 @@ export const updateDistrictsRankings = inngest.createFunction(
           REDIS_KEYS.DISTRICT_REFERRALS_RANKING,
         )
         const results = await Promise.all(
-          districtsReferralsCount.map(entry => upsertDistrictReferralsRanking(entry)),
+          referralsCountByDistrictEntries.map(entry => upsertDistrictReferralsRanking(entry)),
         )
 
         const successfulResults = results.filter(result => result.success)
         const failedResults = results.filter(result => !result.success)
 
         return {
-          totalResults: results.length,
-          successfulResults: successfulResults.length,
-          failedResults: failedResults.length,
+          totalEntries: results.length,
+          successfulEntries: successfulResults.map(result => result.entry),
+          successful: successfulResults.length,
           failedEntries: failedResults.map(result => result.entry),
+          failed: failedResults.length,
         }
       }),
     ])
 
     return {
-      districtAdvocatesCountEntries: districtAdvocatesCounts.length,
-      districtAdvocatesRankingResults,
-      districtReferralsCountEntries: districtsReferralsCount.length,
-      districtReferralsRankingResults,
+      advocatesCountsByDistrictEntries: advocatesCountsByDistrictEntries.length,
+      advocatesRankingResults,
+      referralsCountByDistrictEntries: referralsCountByDistrictEntries.length,
+      referralsRankingResults,
     }
   },
 )
-
-async function fetchStatesCount({
-  queryFn,
-  logger,
-}: {
-  queryFn: (stateCode: USStateCode) => Promise<AdvocatesCountResult[]>
-  logger: Logger
-}) {
-  const states = Object.keys(US_STATE_CODE_TO_DISTRICT_COUNT_MAP) as USStateCode[]
-  const promises = states.map(stateCode =>
-    pRetry(() => queryFn(stateCode), {
-      retries: 3,
-    }),
-  )
-
-  const results = await Promise.allSettled(promises)
-  const allCounts: AdvocatesCountResult[] = []
-  const failedStates: USStateCode[] = []
-
-  results.forEach((result, index) => {
-    const stateCode = states[index]
-    if (result.status === 'fulfilled') {
-      allCounts.push(...result.value)
-    } else {
-      logger.error(`Promise failed for state ${stateCode} in ${queryFn.name}`, {
-        reason: result.reason,
-      })
-      failedStates.push(stateCode)
-    }
-  })
-
-  logger.info(
-    `Fetched ${queryFn.name} counts. Success: ${states.length - failedStates.length}, Failed: ${failedStates.length}`,
-  )
-  return allCounts
-}