closes #2243

fixes <!-- SENTRY issue: Adding the sentry issue number here (e.g.: PROD-SWC-WEB-1JE) will auto-close the issue when this PR is merged. Please ensure sentry issues are marked resolved after we ship related bug fixes -->

## What changed? Why?

**Problem:**
The queries fetching advocate and referral counts by district were extremely slow, taking ~140 seconds to finish and sometimes hitting the 3 minutes limit for edge functions. Initial attempts to optimize the single SQL query through indexing, join order hints, and clause simplification failed to significantly improve performance due to the large data volume involved in joins and aggregations.

**Solution:**
Refactored the data fetching approach to run queries in parallel on a per-state basis:

1.  Modified `getAdvocatesCountByDistrict` and `getReferralsCountByDistrict` to accept a `stateCode` and query data only for that specific state.
    - Leverages the `[administrativeAreaLevel1, countryCode]` index, but only reduced query time by ~3 seconds.
2.  Updated the `updateDistrictsRankings` Inngest job.
    - Query each state count in parallel, greatly reducing total query time.

**Rationale & Alternatives Considered:**

*   **Single Query Optimization:** Attempting to optimize the original query proved insufficient. Despite indexing, query restructuring, and optimizer hints, the inherent complexity of joining and aggregating/grouping across the large `user_action`, `user`, and `address` tables yielded no tangible improvements.
*   **App-Level Grouping:** Rejected as it would require transferring massive amounts of data from the DB to the application, causing network bottlenecks and likely exceeding application memory/CPU limits.
*   **Per-District Queries:** Rejected due to the high overhead of network latency and query planning for hundreds of individual database calls.

The chosen per-state parallel query approach effectively breaks down the work, allowing the database to handle smaller, faster queries concurrently.

**Impact:**

Reduces the total time for fetching district ranking data from ~140 seconds to **~7 seconds**, based on benchmark results. Makes the `updateDistrictsRankings` job significantly faster.

<img width="666" alt="image" src="https://github.com/user-attachments/assets/e1b194f1-86fc-4efd-b355-3308808e0509" />
 
 
<img width="664" alt="image" src="https://github.com/user-attachments/assets/1c8930c8-b537-489a-a568-e4146497537f" />

- Execution Plan Confirmation: `EXPLAIN ANALYZE` output on the refactored per-state queries confirms the database now efficiently uses the `address_administrative_area_level_1_country_code_idx` index for the initial filtering.

<img width="435" alt="image" src="https://github.com/user-attachments/assets/46ce85c1-2b31-40af-95ea-25edc66eec25" />

<img width="549" alt="image" src="https://github.com/user-attachments/assets/e1715af0-7ea3-4797-8890-2f55d3e1abb3" />




<!--
Here you can add any additional context not already captured in related github/sentry issue.
If there are UX changes please do one or both of the following:
    - include paths/steps to reproduce the UI related changes in your vercel preview branch (if you are a core contributor)
    - attach mobile/web screenshots to the PR
-->

## PlanetScale deploy request

<!-- See "Updating the PlanetScale schema" section in docs/Contributing.md -->

## Notes to reviewers

<!-- Here’s where you can give brief guidance on how to review the PR.
(Often it’s helpful to tell reviewers where the “main change” of the PR can be found,
if other diffs in the PR are “ripples” caused by it.)
You can also highlight anything to which you’d like to draw reviewers’ attention. -->

## How has it been tested?

- [ ] Locally
- [ ] Vercel Preview Branch
- [ ] Unit test
- [ ] Functional test